requirements.txt
# requirements.txt

# General libraries
requests==2.28.1
PyPDF2==1.26.0
markdown==3.3.7
python-dotenv==0.20.0

# Libraries for arXiv API
arxiv==0.4.0

# Libraries for LangChain and OpenAI integration
langchain==0.0.200

# Additional libraries for handling PDF generation
xhtml2pdf==0.2.4